requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2-binary==2.9.10
requests==2.32.3
